% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exmplData1}
\alias{exmplData1}
\title{Example dataset1}
\format{
A list with 2 elements:
\describe{
\item{pl1}{97 randomly selected Homo sapiens (Human) proteins randomly selected from UniProt.}
\item{pl2}{45 randomly selected Homo sapiens (Human) proteins randomly selected from UniProt.}
...
}
}
\source{
\url{http://www.uniprot.org/}
}
\usage{
exmplData1
}
\description{
A dataset with randomly selected proteins from UniProt.
}
\keyword{datasets}
