% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exmplData2}
\alias{exmplData2}
\title{Example dataset 2}
\format{
A dataframe with 209 rows and 2 columns.
\describe{
\item{UniProtAC}{UniProt accession code of proteins}
\item{Score}{Check with MJ}
...
}
}
\source{
\doi{molpharm.120.000210}
}
\usage{
exmplData2
}
\description{
Proteins of rat hippocampus proteome.
}
\keyword{datasets}
