% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_ont}
\alias{mod_ont}
\title{Database of protein modifications}
\format{
A data frame with 2102 rows and 3 variables
}
\source{
\url{https://raw.githubusercontent.com/HUPO-PSI/psi-mod-CV/master/PSI-MOD.obo}
}
\usage{
data(mod_ont)
}
\description{
Ontology database for post-translational modification terms. For more details, see the reference.
}
\details{
\itemize{
\item id
\item name
\item def
}
}
\keyword{datasets}
