% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEIMAN-package.R
\docType{package}
\name{PEIMAN-package}
\alias{PEIMAN}
\alias{PEIMAN-package}
\title{PEIMAN: Post-translationsl, Enrichment, Integration, and Matching ANalysis}
\description{
Functions and mined database from UniProt for single enrichment analysis (SEA) and protein set enrichment analysis (PSEA) of a list of protein.
}
\author{
\strong{Maintainer}: Payman Nickchi \email{pnickchi@sfu.ca}

Authors:
\itemize{
  \item Mohieddin Jafari \email{mohieddin.jafari@helsinki.fi}
}

}
\keyword{internal}
