% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEIMAN-package.R
\docType{package}
\name{PEIMAN2-package}
\alias{PEIMAN2}
\alias{PEIMAN2-package}
\title{PEIMAN2: Post-Translational Modification Enrichment, Integration, and Matching Analysis}
\description{
Functions and mined database from 'UniProt' focusing on post-translational modifications to do single enrichment analysis (SEA) and protein set enrichment analysis (PSEA). Payman Nickchi, Mehdi Mirzaie, Marc Baumann, Amir Ata Saei, Mohieddin Jafari (2022) <bioRxiv:10.1101/2022.11.09.515610>.
}
\author{
\strong{Maintainer}: Payman Nickchi \email{pnickchi@sfu.ca}

Authors:
\itemize{
  \item Mohieddin Jafari \email{mohieddin.jafari@helsinki.fi}
}

}
\keyword{internal}
