% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptmlist}
\alias{ptmlist}
\title{Controlled vocabulary of posttranslational modifications (PTM)}
\format{
A data frame with 686 rows and 5 variables
}
\source{
\url{https://www.uniprot.org/docs/ptmlist.txt}
}
\usage{
data(ptmlist)
}
\description{
A dataset containing posttranslational modification information. The variables are as follows:
}
\details{
\itemize{
\item ID Identifier (FT description)
\item AC Accession (PTM-xxxx)
\item KW Keyword
\item FT Feature key
\item DR Cross-reference to external databases
}
}
\keyword{datasets}
