% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exmplData2}
\alias{exmplData2}
\title{Example dataset 2}
\format{
A data frame with 209 rows and 2 columns:
\describe{
\item{UniProtAC}{UniProt accession code of proteins}
\item{Score}{SEQUEST HT score of one associated PSM (used for demonstration purposes)}
...
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/33632781/}
}
\usage{
exmplData2
}
\description{
A test dataset of proteins identified from rat hippocampus
proteome using label-free thermal proteome profiling. The score for each
protein corresponds to the SEQUEST HT engine score of one arbitrary
peptide-spectrum match (PSM) associated with that protein. This dataset is
provided to demonstrate how a ranked list of proteins can be used within the
PEIMAN2 package.
}
\details{
Proteins of rat hippocampus proteome.
}
\keyword{datasets}
