% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getTaxonomyName}
\alias{getTaxonomyName}
\title{Return the exact taxonomy name for list of protein}
\usage{
getTaxonomyName(x)
}
\arguments{
\item{x}{A character vector with each entry presenting a protein UniProt accession code.}
}
\value{
The exact taxonomy name
}
\description{
\code{getTaxonomyName} get a character vector of proteins with their UniProt accession code and returns
the exact taxonomy code.
}
\examples{
getTaxonomyName(x = exmplData1$pl1)
}
