% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEIMAN-package.R
\docType{package}
\name{PEIMAN2-package}
\alias{PEIMAN2}
\alias{PEIMAN2-package}
\title{PEIMAN2: Post-Translational Modification Enrichment, Integration, and Matching Analysis}
\description{
Functions and mined database from 'UniProt' focusing on post-translational modifications to do single enrichment analysis (SEA) and protein set enrichment analysis (PSEA). Payman Nickchi, Uladzislau Vadadokhau, Mehdi Mirzaie, Marc Baumann, Amir Ata Saei, Mohieddin Jafari (2025) \doi{10.1002/pmic.202400238}.
}
\author{
\strong{Maintainer}: Payman Nickchi \email{payman.nickchi@gmail.com}

Authors:
\itemize{
  \item Mohieddin Jafari \email{mohieddin.jafari@helsinki.fi}
}

}
\keyword{internal}
