% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptmlist}
\alias{ptmlist}
\title{Controlled vocabulary for post-translational modifications (PTM) terms}
\format{
A data frame with 686 rows and 5 variables
}
\source{
\url{https://www.uniprot.org/docs/ptmlist.txt}
}
\usage{
data(ptmlist)
}
\description{
This dataframe lists the posttranslational modifications used in the UniProt knowledgebase (Swiss-Prot and TrEMBL).
The columns in this dataframe are as follows:
}
\details{
\itemize{
\item ID Identifier (FT description)
\item AC Accession (PTM-xxxx)
\item KW Keyword
\item FT Feature key
\item DR Cross-reference to external databases
}
}
\keyword{datasets}
